requirements.txt
python-social-auth
beautifulsoup4
django-recaptcha
